@@ -9,6 +9,7 @@
     "url": "https://github.com/nader-eloshaiker"
   },
   "scripts": {
+    "preinstall": "npx only-allow pnpm",
     "dev": "vite --mode dev",
     "build": "tsc && vite build --mode dev",
     "build:prod": "tsc && vite build --mode prod",
@@ -27,75 +28,66 @@
     "@hookform/resolvers": "^3.1.1",
     "@react-hook/resize-observer": "^1.2.6",
     "@tanstack/react-query": "^4.33.0",
-    "axios": "^1.5.1",
+    "axios": "^1.6.0",
     "axios-mock-adapter": "^1.22.0",
     "classnames": "^2.3.2",
-    "daisyui": "^3.9.2",
     "localforage": "^1.10.0",
+    "path-to-regexp": "^6.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.45.1",
+    "react-hook-form": "^7.48.2",
     "react-router-dom": "^6.16.0",
     "simpler-color": "^1.0.1",
-    "styled-components": "^6.1.0",
+    "styled-components": "^6.1.1",
     "vite-plugin-checker": "^0.6.2",
-    "web-vitals": "^2.1.4",
+    "web-vitals": "^3.5.0",
     "yup": "^1.2.0"
   },
   "devDependencies": {
-    "@tailwindcss/forms": "^0.5.3",
-    "@tailwindcss/typography": "^0.5.9",
     "@testing-library/jest-dom": "^6.1.2",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
-    "@types/autoprefixer": "^10.2.0",
-    "@types/axios": "^0.14.0",
-    "@types/axios-mock-adapter": "^1.10.0",
-    "@types/classnames": "^2.3.1",
-    "@types/eslint-config-prettier": "^6.11.0",
+    "@types/eslint-config-prettier": "^6.11.2",
     "@types/eslint-plugin-prettier": "^3.1.2",
     "@types/jsdom": "^21.1.2",
     "@types/lint-staged": "^13.2.0",
-    "@types/localforage": "^0.0.34",
     "@types/node": "^16.18.36",
-    "@types/prettier": "^3.0.0",
-    "@types/react": "^18.2.14",
-    "@types/react-dom": "^18.2.6",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
     "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.26",
-    "@types/tailwindcss": "^3.1.0",
-    "@types/yup": "^0.32.0",
-    "@typescript-eslint/eslint-plugin": "^5.60.0",
-    "@typescript-eslint/parser": "^5.60.0",
-    "@vitejs/plugin-react": "^4.0.4",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "@vitejs/plugin-react": "^4.1.1",
     "@vitest/coverage-v8": "^0.34.6",
     "@vitest/ui": "^0.34.6",
     "autoprefixer": "^10.4.16",
+    "daisyui": "^4.1.0",
     "eslint": "^8.48.0",
-    "eslint-config-prettier": "^8.8.0",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^5.0.1",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.1",
+    "eslint-plugin-react-refresh": "^0.4.3",
     "eslint-plugin-tailwindcss": "^3.13.0",
     "husky": "^8.0.3",
     "jsdom": "^22.1.0",
-    "lint-staged": "^15.0.2",
+    "lint-staged": "^15.1.0",
     "orval": "^6.17.0",
-    "postcss": "^8.4.31",
     "prettier": "^3.0.3",
-    "prettier-plugin-organize-imports": "^3.2.2",
-    "react-scripts": "^5.0.1",
-    "tailwind-merge": "^1.13.2",
-    "tailwindcss": "^3.3.3",
-    "typescript": "^4.9.5",
-    "vite": "^4.3.9",
+    "prettier-plugin-organize-imports": "^3.2.3",
+    "tailwind-merge": "^2.0.0",
+    "tailwindcss": "^3.3.5",
+    "typescript": "^5.0.2",
+    "vite": "^4.5.0",
     "vite-plugin-checker": "^0.6.2",
-    "vite-tsconfig-paths": "^4.2.0",
     "vitest": "^0.34.6"
   },
+  "browserslist": [
+    "defaults"
+  ],
   "lint-staged": {
     "**/*.{ts,tsx,js} !(src/generated/**/*)": "npm run lint:fix",
     "*.{ts,tsx,js,css,md*}": "npm run format"
