@@ -1,13 +1,11 @@
 /** @type {import('tailwindcss').Config} */
 
-import forms from '@tailwindcss/forms'
-import typograhpy from '@tailwindcss/typography'
 import daisyui from 'daisyui'
-import daisyuiThemes from 'daisyui/src/theming/themes'
+import { themes as daisyuiThemes } from 'daisyui/src/theming/themes'
 import colors from 'tailwindcss/colors'
 import defaultTheme from 'tailwindcss/defaultTheme'
 
-module.exports = {
+export default {
   mode: 'jit',
   content: ['./src/index.html', './src/**/*.{js,ts,jsx,tsx,html}'],
   theme: {
@@ -36,37 +34,50 @@ module.exports = {
       },
     },
   },
-  plugins: [forms, typograhpy, daisyui],
+  plugins: [daisyui],
   darkMode: ['class', '[data-theme="dark"]'],
   daisyui: {
     themes: [
       {
         light: {
-          ...daisyuiThemes['[data-theme=light]'],
-          '.btn-action': {
-            color: '#d1620d',
-            'border-color': '#d1620d',
-          },
-          '.btn-action:hover': {
-            color: '#fde68a',
-            'border-color': '#d1620d',
-            'background-color': '#d1620d',
-          },
+          ...daisyuiThemes['light'],
+          // 'color-scheme': 'light',
+          primary: '#e37c07',
+          'primary-content': '#f5ecdc',
+          secondary: '#0483d2',
+          'secondary-content': '#c6e4f7',
+          accent: '#e600cf',
+          'accent-content': '#fcc0f6',
+          // neutral: '#2B3440',
+          // 'neutral-content': '#D7DDE4',
+          // 'base-100': '#ffffff',
+          // 'base-200': '#F2F2F2',
+          // 'base-300': '#E5E6E6',
+          // 'base-content': '#1f2937',
         },
       },
       'dark',
       {
         dark: {
-          ...daisyuiThemes['[data-theme=dark]'],
-          '.btn-action': {
-            color: '#d1620d',
-            'border-color': '#d1620d',
-          },
-          '.btn-action:hover': {
-            color: '#fde68a',
-            'border-color': '#d1620d',
-            'background-color': '#d1620d',
-          },
+          ...daisyuiThemes['dark'],
+          // 'color-scheme': 'dark',
+          primary: '#cc5b14',
+          'primary-content': '#f5ecdc',
+          secondary: '#059cfa',
+          'secondary-content': '#d7eefc',
+          accent: '#ff0fe8',
+          'accent-content': '#ffd4fb',
+          // neutral: '#35253c',
+          // 'neutral-focus': '#242b33',
+          // 'neutral-content': '#A6ADBB',
+          // 'base-100': '#1c1917',
+          // 'base-200': '#292524',
+          // 'base-300': '#44403c',
+          // 'base-content': '#A6ADBB',
+          // info: '#419ec3',
+          // success: '#19cca2',
+          // warning: '#fbb565',
+          // error: '#fc5985',
         },
       },
     ],
