@@ -1,6 +1,5 @@
-import { Params } from 'react-router-dom'
-import { routes } from '../../routes/AppRouteSchema'
-import { ScreenInput, ScreenInputList, ScreenItem } from './../../generated/openapi/models'
+import { ScreenInput, ScreenInputList, ScreenItem } from '../../generated/openapi/models'
+import { transformScreenInput } from '../../utils/ScreenTransformation'
 import { createItem, createItemList, deleteItem, getItem, getItemList, updateItem } from './indexDB'
 
 export type TScreenListResponse = {
@@ -21,8 +20,11 @@ export async function getScreenList(): Promise<TScreenListResponse> {
   return { list }
 }
 
-export async function getScreen(params: Params): Promise<{ item: ScreenItem | undefined }> {
-  const item = await getItem(params[routes.screens.key])
+export async function getScreen(id: string | undefined): Promise<{ item: ScreenItem | undefined }> {
+  if (!id) {
+    throw new Error('No id provided!')
+  }
+  const item = await getItem(id)
   if (!item) {
     throw new Error('Not Found')
   }
@@ -30,8 +32,12 @@ export async function getScreen(params: Params): Promise<{ item: ScreenItem | un
   return { item }
 }
 
-export async function updateItemAction(data: ScreenItem) {
-  const item = await updateItem(data.id || '', data)
+export async function updateItemAction(id: string | undefined, data: ScreenInput) {
+  if (!id) {
+    throw new Error('No id provided!')
+  }
+  const screenItem = transformScreenInput(data, id)
+  const item = await updateItem(id, screenItem)
 
   return { item }
 }
@@ -48,13 +54,7 @@ export async function createItemListAction(data: ScreenInputList) {
   return { list }
 }
 
-export async function deleteItemAction(url: string | undefined) {
-  if (!url) {
-    throw new Error('No url provided!')
-  }
-
-  const id = url.split('/').at(-1) ?? ''
-
+export async function deleteItemAction(id: string | undefined) {
   if (!id) {
     throw new Error('No id provided!')
   }
@@ -64,13 +64,7 @@ export async function deleteItemAction(url: string | undefined) {
   return { id }
 }
 
-export async function favouriteItemAction(url: string | undefined) {
-  if (!url) {
-    throw new Error('No url provided!')
-  }
-
-  const id = url.split('/').at(-2) ?? ''
-
+export async function favouriteItemAction(id: string | undefined) {
   if (!id) {
     throw new Error('No id provided!')
   }
