@@ -1,10 +1,11 @@
 // ./components/Autocomplete.tsx
 
 import cn from 'classnames'
-import { ChangeEvent, memo, useRef, useState } from 'react'
+import { ChangeEvent, KeyboardEvent, memo, useRef, useState } from 'react'
 import MagnifyGlassIcon from '../../assets/icons/MagnifyGlass'
 import { useElementSize } from '../../hooks/useElementSize'
-import { InputPlaceholder } from '../inputplaceholder/InputPlaceholder'
+import { InputFix } from '../input-suffix/InputFix'
+import { FixLocation } from '../input-suffix/InputFix.type'
 
 export type TAutoCompleteItem = { id: string; label: string }
 type TProps = TRestProps & {
@@ -30,17 +31,14 @@ export const AutoComplete = ({
 }: TProps) => {
   const divRef = useRef<HTMLDivElement>(null)
   const { width } = useElementSize(divRef)
-  // const { ref: divRef, width = 1 } = useResizeObserver<HTMLDivElement>()
   const [open, setOpen] = useState(false)
-  const [inputValue, setInputValue] = useState(items.length === 1 ? items[0].label : value)
 
-  const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
-    onChange(event.target.value)
-    setInputValue(event.target.value)
+  const handleChange = ({ target }: ChangeEvent<HTMLInputElement>) => {
+    onChange(target.value)
   }
 
   const handleSelect = (item: TAutoCompleteItem) => {
-    setInputValue(item.label)
+    onChange(item.label)
     onSelect(item)
     setOpen(false)
     // force close of the dropdown
@@ -49,41 +47,49 @@ export const AutoComplete = ({
     }
   }
 
+  const handleKeyDown = (event: KeyboardEvent<HTMLLIElement>, item: TAutoCompleteItem) => {
+    if (event.code === 'Enter') {
+      handleSelect(item)
+    }
+  }
+
   return (
     <div
       data-testid='autoComplete'
       // use classnames here to easily toggle dropdown open
-      className={
-        cn({
-          'dropdown w-full relative': true,
-          'dropdown-open': open,
-        }) + (className ? ' ' + className : '')
-      }
+      className={cn(className, {
+        'dropdown w-full': true,
+        'dropdown-open': open,
+      })}
       ref={divRef}
       {...rest}
     >
-      <InputPlaceholder data-testid='autoCompleteLoading' className='absolute z-10 flex h-full w-8 items-center pl-2'>
-        {isLoading ? (
-          <span className='loading loading-spinner loading-md' />
-        ) : (
-          <MagnifyGlassIcon className='flex h-6 w-6' />
-        )}
-      </InputPlaceholder>
-      <input
-        name='autoCompleteInput'
-        data-testid='autoCompleteInput'
-        type='text'
-        className={cn({ relative: isLoading }) + ' input input-md w-full pl-10 relative shadow-md'}
-        value={inputValue}
-        onChange={handleChange}
-        placeholder={placeholder || 'Type something..'}
-        tabIndex={0}
-        disabled={isLoading}
-      />
+      <InputFix
+        fix={
+          isLoading ? (
+            <span data-testid='autoCompleteLoading' className='loading loading-spinner loading-md' />
+          ) : (
+            <MagnifyGlassIcon data-testid='autoCompleteIcon' className='h-6 w-6' />
+          )
+        }
+        location={FixLocation.prefix}
+      >
+        <input
+          name='autoCompleteInput'
+          data-testid='autoCompleteInput'
+          type='text'
+          className='input input-md relative w-full !pl-12 shadow-md'
+          value={value}
+          onChange={handleChange}
+          placeholder={placeholder ?? 'Type something..'}
+          tabIndex={0}
+          disabled={isLoading}
+        />
+      </InputFix>
       {items.length > 0 && (
         <div
           data-testid='autoCompleteResults'
-          className='dropdown-content top-14 z-40 max-h-80 flex-col overflow-auto rounded-md bg-base-200'
+          className='dropdown-content top-14 z-[1] max-h-80 flex-col overflow-auto rounded-md bg-base-200'
         >
           <ul
             className='menu'
@@ -96,6 +102,7 @@ export const AutoComplete = ({
                   key={item.id}
                   tabIndex={index + 1}
                   onClick={() => handleSelect(item)}
+                  onKeyDown={(event) => handleKeyDown(event, item)}
                   className='w-full border-b border-b-base-content/10'
                 >
                   <button>{item.label}</button>
