@@ -3,6 +3,8 @@ import CloseIcon from '../../assets/icons/Close'
 import EditIcon from '../../assets/icons/Edit'
 import StarOutlineIcon from '../../assets/icons/StarOutline'
 import StarSolidIcon from '../../assets/icons/StarSolid'
+import { FormDrawerActionTypes } from '../../contexts/FormDrawer/FormDrawerManager'
+import { useFormDrawerContext } from '../../contexts/FormDrawer/useFormDrawaerContext'
 import { ScreenItem } from '../../generated/openapi/models'
 import { useDeleteScreen } from '../../hooks/api/useDeleteScreen'
 import { useFavoriteScreen } from '../../hooks/api/useFavoriteScreen'
@@ -51,6 +53,7 @@ export const ScreenTable = ({
 }: Props) => {
   const { isDeleteLoading, deleteAction } = useDeleteScreen()
   const { isFavoriteLoading, favoriteAction } = useFavoriteScreen()
+  const { dispatchFormDrawer } = useFormDrawerContext()
   const [selected, setSelected] = useState<ScreenItem>()
   const [themeMode] = useThemeMode()
 
@@ -62,6 +65,12 @@ export const ScreenTable = ({
   const handleDelete = (screen: ScreenItem) => {
     setSelected(screen)
     deleteAction({ id: screen.id })
+  }
+
+  const handleEdit = (screen: ScreenItem) => {
+    setSelected(screen)
+
+    dispatchFormDrawer({ type: FormDrawerActionTypes.Edit, payload: { id: screen.id } })
   }
 
   return (
@@ -96,6 +105,7 @@ export const ScreenTable = ({
               key={screen.id}
               onMouseEnter={() => setHighLighted(screen)}
               onMouseOut={() => setHighLighted(undefined)}
+              onBlur={() => setHighLighted(undefined)}
               onClick={() => onHighlightClick(screen)}
             >
               <td>
@@ -136,7 +146,7 @@ export const ScreenTable = ({
               <td className='text-center'>{screen.spec && `${Math.round((screen.spec.ppi * 100) / 100)}`}</td>
               <td>
                 <div className='flex flex-row items-center justify-center gap-3'>
-                  <button>
+                  <button onClick={() => handleEdit(screen)}>
                     <EditIcon id='edit-icon' className='h-4 w-4' fill='currentColor' />
                   </button>
                   {isDeleteLoading && screen.id === selected?.id ? (
