@@ -6,10 +6,13 @@ type TProps = TRestProps & { className?: string }
 export const SkeletonImage = ({ className, ...rest }: TProps) => {
   return (
     <div
-      className={twMerge(className, 'flex animate-pulse items-center justify-center rounded-md border-2 bg-gray-300')}
+      className={twMerge(
+        className,
+        'flex items-center justify-center rounded-md skeleton bg-neutral-300 dark:bg-neutral-600 ',
+      )}
       {...rest}
     >
-      <ImageIcon className='h-10 w-10' fill='rgb(107 114 128)' />
+      <ImageIcon className='h-20 w-20' fill='rgb(107 114 128)' />
     </div>
   )
 }
