@@ -6,47 +6,47 @@ export const defaultScreenInputList: ScreenInputList = [
     hRes: 3440,
     vRes: 1440,
     aspectRatio: '21:9',
-    lightColor: '#68CA68',
-    darkColor: '#359735',
+    lightColor: '#91A68C',
+    darkColor: '#5E7359',
   },
   {
     diagonalSize: 38,
     hRes: 3840,
     vRes: 1600,
     aspectRatio: '21:9',
-    lightColor: '#7CACB6',
-    darkColor: '#497983',
+    lightColor: '#F6693C',
+    darkColor: '#C33609',
   },
   {
     diagonalSize: 27,
     hRes: 3840,
     vRes: 2160,
     aspectRatio: '16:9',
-    lightColor: '#A38FA3',
-    darkColor: '#705C70',
+    lightColor: '#A2DB57',
+    darkColor: '#6FA824',
   },
   {
     diagonalSize: 32,
     hRes: 3840,
     vRes: 2160,
     aspectRatio: '16:9',
-    lightColor: '#999999',
-    darkColor: '#666666',
+    lightColor: '#8EA4A4',
+    darkColor: '#5B7171',
   },
   {
     diagonalSize: 40,
     hRes: 5120,
     vRes: 2160,
     aspectRatio: '21:9',
-    lightColor: '#B979AA',
-    darkColor: '#864677',
+    lightColor: '#D7E052',
+    darkColor: '#A4AD1F',
   },
   {
     diagonalSize: 49,
     hRes: 5120,
     vRes: 1440,
     aspectRatio: '32:9',
-    lightColor: '#A7A78B',
-    darkColor: '#747458',
+    lightColor: '#E74B61',
+    darkColor: '#B4182E',
   },
 ]
