@@ -15,7 +15,7 @@ export * from './screenIdResponse'
 export * from './screenInput'
 export * from './screenInputList'
 export * from './screenItem'
-export * from './screenItemRespose'
+export * from './screenItemResponse'
 export * from './screenListResponse'
 export * from './screenRender'
 export * from './screenSpec'
