@@ -1,3 +1,4 @@
+import { UseMutationOptions } from '@tanstack/react-query'
 import { useEffect } from 'react'
 import {
   GeneralNotificationItem,
@@ -7,13 +8,20 @@ import {
 import { useNotificationContext } from '../../contexts/Notification/useNotifcationContext'
 import { ScreenActionTypes } from '../../contexts/Screen/ScreenManager'
 import { useScreenContext } from '../../contexts/Screen/useScreenContext'
-import { ScreenInput } from '../../generated/openapi/models'
+import { ErrorResponse, ScreenInput, ScreenItemResponse } from '../../generated/openapi/models'
 import { useCreateScreenAction } from '../../generated/openapi/services/screen-list-service'
 
-export const useCreateScreen = () => {
+export type CreateScreenOptions = UseMutationOptions<ScreenItemResponse, ErrorResponse, { data: ScreenInput }>
+
+export const useCreateScreen = (queryOptions?: CreateScreenOptions) => {
   const { dispatch: dispatchScreen } = useScreenContext()
   const { dispatch: dispatchNotification } = useNotificationContext()
-  const { isLoading: isCreateLoading, data: createResponse, error: createError, mutate } = useCreateScreenAction()
+  const {
+    isLoading: isCreateLoading,
+    data: createResponse,
+    error: createError,
+    mutate,
+  } = useCreateScreenAction({ mutation: queryOptions })
 
   useEffect(() => {
     if (createResponse) {
