@@ -1,4 +1,6 @@
+import { UseMutationOptions } from '@tanstack/react-query'
 import { useEffect } from 'react'
+import { ErrorResponse } from 'react-router-dom'
 import {
   GeneralNotificationItem,
   NotificationActionTypes,
@@ -7,9 +9,12 @@ import {
 import { useNotificationContext } from '../../contexts/Notification/useNotifcationContext'
 import { ScreenActionTypes } from '../../contexts/Screen/ScreenManager'
 import { useScreenContext } from '../../contexts/Screen/useScreenContext'
+import { ScreenIdResponse } from '../../generated/openapi/models'
 import { useDeleteScreenAction } from '../../generated/openapi/services/screen-service'
 
-export const useDeleteScreen = () => {
+export type DeleteScreenOptions = UseMutationOptions<ScreenIdResponse, ErrorResponse, { id: string }>
+
+export const useDeleteScreen = (queryOptions?: DeleteScreenOptions) => {
   const { dispatch: dispatchScreen } = useScreenContext()
   const { dispatch: dispatchNotification } = useNotificationContext()
   const {
@@ -17,7 +22,7 @@ export const useDeleteScreen = () => {
     data: deleteResponse,
     error: deleteError,
     mutate: deleteAction,
-  } = useDeleteScreenAction()
+  } = useDeleteScreenAction({ mutation: queryOptions })
 
   useEffect(() => {
     if (deleteResponse) {
