@@ -1,14 +1,23 @@
+import { UseQueryOptions } from '@tanstack/react-query'
 import { useEffect } from 'react'
+import { ErrorResponse } from 'react-router-dom'
 import { NotificationActionTypes, NotificationType } from '../../contexts/Notification/NotificationManager'
 import { useNotificationContext } from '../../contexts/Notification/useNotifcationContext'
 import { ScreenActionTypes } from '../../contexts/Screen/ScreenManager'
 import { useScreenContext } from '../../contexts/Screen/useScreenContext'
+import { ScreenListResponse } from '../../generated/openapi/models'
 import { useListScreensAction } from '../../generated/openapi/services/screen-list-service'
 
-export const useListScreens = () => {
+export type ListScreenOptions = UseQueryOptions<ScreenListResponse, ErrorResponse>
+
+export const useListScreens = (queryOptions?: ListScreenOptions) => {
   const { dispatch: dispatchScreen } = useScreenContext()
   const { dispatch: dispatchNotification } = useNotificationContext()
-  const { isLoading: isScreenListLoading, error: screenListError, data: screenListResponse } = useListScreensAction()
+  const {
+    isFetching: isScreenListLoading,
+    error: screenListError,
+    data: screenListResponse,
+  } = useListScreensAction({ query: queryOptions })
 
   useEffect(() => {
     if (screenListResponse && screenListResponse.list.length > 0) {
