@@ -6,7 +6,7 @@ export default function Root() {
   return (
     <div id='app-root' className='container mx-auto flex h-full min-h-screen w-full flex-col'>
       <Header />
-      <main role='main' className='flex-1 px-2 sm:px-4'>
+      <main role='main' className='my-6 flex-1'>
         <Outlet />
       </main>
       <Footer />
